Move check_lmp_columns into sole ISO test suite using it, test_pjm
